+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/Block/utils/TemplateString/Variable/utils/index.ts
@@ -3,10 +3,9 @@ import {
   VariableObject,
   VariableObjectValidated,
   VariableType,
-} from "../../../../../schema/primitives/variable";
+} from "../../../../../../schema/primitives/variable";
 
-import { TemplateStringSeparator } from "../../../../../schema/primitives/templateString";
-import { useQueryClient } from "@tanstack/react-query";
+import { TemplateStringSeparator } from "../../../../../../schema/primitives/templateString";
 import { z } from "zod";
 
 /**
@@ -124,76 +123,4 @@ export const getValueFromJsonPath = (
   }
 
   return [returnValue, undefined];
-};
-
-type UseVariableSuccess = [PossibleValues, undefined];
-type QueryFailure = [undefined, "queryIdNotFound" | "couldNotStringify"];
-
-type UseVariableFailure =
-  | GetValueFromJsonPathFailure
-  | ValidateVariableFailure
-  | QueryFailure;
-
-export const useVariable = (
-  value: VariableType
-): UseVariableSuccess | UseVariableFailure => {
-  const queryClient = useQueryClient();
-  const [variableObject, validationError] = validateVariable(
-    parseVariable(value)
-  );
-
-  if (validationError) {
-    return [undefined, validationError];
-  }
-
-  const { id, pointer } = variableObject;
-  const cacheKey = [id];
-  const queryState = queryClient.getQueryState(cacheKey);
-
-  if (queryState === undefined) {
-    return [undefined, "queryIdNotFound"];
-  }
-
-  const cachedData = queryClient.getQueryData(cacheKey);
-  const [data, error] = getValueFromJsonPath(cachedData, pointer);
-
-  if (error) {
-    return [undefined, error];
-  }
-
-  return [data, undefined];
-};
-
-export const JSXValueSchema = z.union([
-  z.string(),
-  z.number(),
-  z.boolean(),
-  z.null(),
-  z.undefined(),
-]);
-
-export type JSXValueType = z.infer<typeof JSXValueSchema>;
-
-type UseVariableJSXSuccess = [JSXValueType, undefined];
-
-type UseVariableJSXFailure =
-  | GetValueFromJsonPathFailure
-  | ValidateVariableFailure
-  | QueryFailure;
-
-export const useVariableJSX = (
-  value: VariableType
-): UseVariableJSXSuccess | UseVariableJSXFailure => {
-  const [data, error] = useVariable(value);
-
-  if (error) {
-    return [undefined, error];
-  }
-
-  const dataParsed = JSXValueSchema.safeParse(data);
-  if (!dataParsed.success) {
-    return [undefined, "couldNotStringify"];
-  }
-
-  return [dataParsed.data, undefined];
 };
